+++ noodles-bam/src/async/reader.rs
@@ -1,8 +1,5 @@
-mod builder;
 mod query;
 mod record;
-
-pub use self::builder::Builder;
 
 use bytes::BytesMut;
 use futures::{stream, Stream};
@@ -333,20 +330,6 @@ impl<R> Reader<bgzf::AsyncReader<R>>
 where
     R: AsyncRead + Unpin,
 {
-    /// Creates an async BAM reader builder.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use noodles_bam as bam;
-    /// let data = [];
-    /// let builder = bam::AsyncReader::builder(&data[..]);
-    /// let reader = builder.build();
-    /// ```
-    pub fn builder(inner: R) -> Builder<R> {
-        Builder::new(inner)
-    }
-
     /// Creates an async BAM reader.
     ///
     /// # Examples